+++ python/megane/__init__.py
@@ -1,6 +1,7 @@
 """megane - A fast, beautiful molecular viewer."""
 
 from megane.parsers.cif import load_cif
+from megane.parsers.jcampdx import load_jcampdx
 from megane.parsers.lammps_data import load_lammps_data
 from megane.parsers.lammpstrj import load_lammpstrj_structure
 from megane.parsers.pdb import load_pdb
@@ -15,6 +16,7 @@ from megane.pipeline import (
     Color,
     Filter,
     Isosurface,
+    LoadSpectrum,
     LoadStructure,
     LoadTrajectory,
     LoadVector,
@@ -23,6 +25,7 @@ from megane.pipeline import (
     Pipeline,
     Replicate,
     Representation,
+    SpectrumPlot,
     Streaming,
     VectorOverlay,
     Viewport,
@@ -40,6 +43,7 @@ __all__ = [
     "Filter",
     "build_pipeline",
     "Isosurface",
+    "LoadSpectrum",
     "LoadStructure",
     "LoadTrajectory",
     "LoadVector",
@@ -49,10 +53,12 @@ __all__ = [
     "Pipeline",
     "Replicate",
     "Representation",
+    "SpectrumPlot",
     "Streaming",
     "VectorOverlay",
     "Viewport",
     "load_cif",
+    "load_jcampdx",
     "load_lammps_data",
     "load_lammpstrj_structure",
     "load_pdb",